backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.0.1
bcrypt==5.0.0
PyJWT==2.9.0
mysql-connector-python==9.6.0
